refactor: move config generation to shared
